pyproject.toml
[tool.poetry]
name = "trashy-poetry"
version = "0.4.1"
description = "Poetry for everyday use"
authors = ["Paul Bailey <paul@neutron.studio>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/pizzapanther/prose"
packages = [
    { include = "prose" }
]

[tool.poetry.dependencies]
python = ">=3.6,<4.0"
poetry = "^1.1.4"
python-dotenv = "^0.15.0"
poethepoet = "^0.10.0"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"

[tool.poetry.scripts]
prose = "prose.cli:main"

[tool.prose.env]
TEST_ENV = "hello world"
PATH = "${PATH}:/narf"
PYTHONPATH = "${PROSE_PROJECT_HOME}:${PYTHONPATH}"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poe.tasks]
check_format = "yapf -rpd prose/"
format = "yapf -rpi -vv prose/"
